💄 Code formatting for Java, Shell, JS, YAML too (#1470)
Prettier will format Java, Shell, JS, and YAML.  This will be done in the pre-commit hook and in the Policy CI.
Swift and Kotlin are not production ready.  Ruby is hard to identify.  Dart is done as part of Client CI and `flutter format`.

If something is broken, it is most likely NOT this commit.  Look before it in the `git blame`