Client race condition fixes (putLocation and putDeviceToken in single request ) (#1820)
Client and Server changes to obsolete putDeviceToken and putLocation in favor of new putNotificationSettings. This reduces the chance of a race condition in pushing client settings to the server.

It also adds a tiny bit of debug logging on the server.

The behavior should be mostly the same as before in terms of when RPCs are fired. 

No true locking is present either client or server side at this time, but the requests contain the full client state.

A future change could include server-side transactions.

Addresses issue #1524